backend/requirements.txt
fastapi
uvicorn[standard]
azure-cosmos==4.5.1
requests
pandas
azure-identity
azure-keyvault-secrets
python-multipart
pydantic
jinja2==3.1.4
openai
mlflow
pydantic